youtube/requirements.txt
black==22.10.0
click==8.1.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
python-dateutil==2.8.2
python-decouple==3.6
schedule==1.1.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
